Refactor install command to use guidance module for intent skills block handling
