+++ packages/intent/src/install/guidance.ts
